requirements.txt
streamlit
deepface==0.0.79
opencv-python
Pillow
